backend/requirements.txt
fastapi
uvicorn
pandas
langchain
langchain-community
sentence-transformers
faiss-cpu
pydantic
python-dotenv
